# specs/001-pm-simulation-web/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Olivia PM Simulation API
  version: 0.1.0
servers:
  - url: https://api.example.com
paths:
  /scenarios:
    get:
      summary: List available scenarios
      responses:
        '200':
          description: Scenario catalog grouped by discipline (PM, PMO)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Scenario'
  /scenarios/{id}:
    get:
      summary: Get scenario detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Scenario detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
  /sessions:
    post:
      summary: Create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionWithKickoff'
    get:
      summary: List sessions
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, completed, evaluated]
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Sessions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistoryItem'
  /sessions/{id}:
    get:
      summary: Get session detail
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session with transcript/actions/evaluation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
    delete:
      summary: Delete or archive session
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '204':
          description: Deleted
  /sessions/{id}/messages:
    post:
      summary: Send user message and receive agent reply
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
      responses:
        '200':
          description: Agent reply with updated session state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /sessions/{id}/evaluate:
    post:
      summary: Trigger evaluation
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Evaluation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evaluation'
  /sessions/{id}/comments:
    post:
      summary: Add a manager comment to a session
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagerComment'
    get:
      summary: List manager comments for a session
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Comments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManagerComment'

components:
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    CreateSessionRequest:
      type: object
      required: [scenarioId]
      properties:
        scenarioId:
          type: string
    CreateMessageRequest:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user]
        content:
          type: string
        tags:
          type: array
          items:
            type: string
            enum: [decision, assumption, risk, next_action, summary]
    CreateCommentRequest:
      type: object
      required: [content]
      properties:
        authorName:
          type: string
          nullable: true
        content:
          type: string
    SessionWithKickoff:
      type: object
      properties:
        session:
          $ref: '#/components/schemas/Session'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
    SessionDetail:
      type: object
      properties:
        session:
          $ref: '#/components/schemas/Session'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        evaluation:
          $ref: '#/components/schemas/Evaluation'
        actions:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/ManagerComment'
    MessageResponse:
      type: object
      properties:
        reply:
          $ref: '#/components/schemas/Message'
        session:
          $ref: '#/components/schemas/Session'
    Session:
      type: object
      properties:
        id:
          type: string
        scenarioId:
          type: string
        scenarioDiscipline:
          type: string
          enum: [PM, PMO]
        status:
          type: string
          enum: [active, completed, evaluated]
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        lastActivityAt:
          type: string
          format: date-time
        userName:
          type: string
          nullable: true
        progressFlags:
          type: object
          properties:
            requirements:
              type: boolean
            priorities:
              type: boolean
            risks:
              type: boolean
            acceptance:
              type: boolean
        evaluationRequested:
          type: boolean
        missionStatus:
          type: array
          items:
            $ref: '#/components/schemas/MissionStatus'
    Message:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        role:
          type: string
          enum: [user, agent, system]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
            enum: [decision, assumption, risk, next_action, summary]
        queuedOffline:
          type: boolean
    Evaluation:
      type: object
      properties:
        sessionId:
          type: string
        overallScore:
          type: number
          format: float
        passing:
          type: boolean
        categories:
          type: array
          items:
            $ref: '#/components/schemas/EvaluationCategory'
        summary:
          type: string
        improvementAdvice:
          type: string
    EvaluationCategory:
      type: object
      properties:
        name:
          type: string
        weight:
          type: number
          format: float
        score:
          type: number
          format: float
        feedback:
          type: string
    HistoryItem:
      type: object
      properties:
        sessionId:
          type: string
        scenarioId:
          type: string
        scenarioDiscipline:
          type: string
          enum: [PM, PMO]
        metadata:
          type: object
          properties:
            duration:
              type: number
            messageCount:
              type: integer
        actions:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        evaluation:
          $ref: '#/components/schemas/Evaluation'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/ManagerComment'
    Scenario:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        discipline:
          type: string
          enum: [PM, PMO]
        description:
          type: string
        product:
          $ref: '#/components/schemas/ScenarioProduct'
        mode:
          type: string
          enum: [freeform]
        kickoffPrompt:
          type: string
        evaluationCriteria:
          type: array
          items:
            $ref: '#/components/schemas/EvaluationCategory'
        passingScore:
          type: number
          format: float
        missions:
          type: array
          items:
            $ref: '#/components/schemas/Mission'
        supplementalInfo:
          type: string
          description: Markdown/plain text for Scenario side info
    ScenarioProduct:
      type: object
      properties:
        name:
          type: string
        summary:
          type: string
        audience:
          type: string
        problems:
          type: array
          items:
            type: string
        goals:
          type: array
          items:
            type: string
        differentiators:
          type: array
          items:
            type: string
        scope:
          type: array
          items:
            type: string
        constraints:
          type: array
          items:
            type: string
        timeline:
          type: string
        successCriteria:
          type: array
          items:
            type: string
        uniqueEdge:
          type: string
          nullable: true
        techStack:
          type: array
          items:
            type: string
        coreFeatures:
          type: array
          items:
            type: string
    Mission:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        order:
          type: integer
    MissionStatus:
      type: object
      properties:
        missionId:
          type: string
        completedAt:
          type: string
          format: date-time
          nullable: true
    ManagerComment:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        authorName:
          type: string
          nullable: true
        content:
          type: string
        createdAt:
          type: string
          format: date-time